+++ src/Giraffe/Routing.fs
@@ -213,3 +213,43 @@ let subRoute (path : string) (handler : HttpHandler) : HttpHandler =
 let subRouteCi (path : string) (handler : HttpHandler) : HttpHandler =
     routeStartsWithCi path >=>
     handlerWithRootedPath path handler
+
+/// ** Description **
+/// Filters an incoming HTTP request based on a part of the request path (case sensitive).
+/// If the sub route matches the incoming HTTP request then the arguments from the `PrintfFormat<...>` will be automatically resolved and passed into the supplied `routeHandler`.
+///
+/// ** Supported format chars **
+///     - `%b`: `bool`
+///     - `%c`: `char`
+///     - `%s`: `string`
+///     - `%i`: `int`
+///     - `%d`: `int64`
+///     - `%f`: `float`/`double`
+///     - `%O`: `Guid`
+///
+/// Subsequent routing handlers inside the given `handler` function should omit the already validated path.
+///
+/// ** Parameters **
+///     - `path`: A format string representing the expected request sub path.
+///     - `routeHandler`: A function which accepts a tuple `'T` of the parsed arguments and returns a `HttpHandler` function which will subsequently deal with the request.
+///
+/// ** Output **
+/// A Giraffe `HttpHandler` function which can be composed into a bigger web application.
+let subRoutef (path : PrintfFormat<_,_,_,_, 'T>) (routeHandler : 'T -> HttpHandler) : HttpHandler =
+        validateFormat path
+        fun (next : HttpFunc) (ctx : HttpContext) ->
+            let paramCount   = (path.Value.Split '/').Length
+            let subPathParts = (getPath ctx).Split '/'
+            if paramCount > subPathParts.Length then abort
+            else
+                let subPath =
+                    subPathParts
+                    |> Array.take paramCount
+                    |> Array.fold (fun state elem ->
+                        if String.IsNullOrEmpty elem
+                        then state
+                        else sprintf "%s/%s" state elem) ""
+                tryMatchInput path subPath false
+                |> function
+                    | None      -> abort
+                    | Some args -> handlerWithRootedPath subPath (routeHandler args) next ctx